feat(log): provide simple console reporter
The simple console reporter will be used if `standard-log-console` is not available

Moving code from `standard-log-console` to `standard-log`.
`standard-log-console` will do more fancy stuff like color.

In the process of eliminating `standard-log-core` so that
`standard-log` will be the only place to keep state.

This will enable more secure features to be added